backend/requirements.txt
# Core web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Data validation and settings
pydantic>=2.10.0
pydantic-settings==2.7.0

# HTTP clients
httpx==0.28.1
requests==2.32.3

# JSON processing
orjson==3.10.12

# Vector database and embeddings
chromadb==1.5.1
sentence-transformers==3.3.1
huggingface_hub==0.27.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0

# Configuration
python-dotenv==1.0.1

# LangChain ecosystem (Ollama integration)
langchain==0.3.15
langchain-ollama==0.3.0
langchain-core>=0.3.47
langchain-community==0.3.15

# Caching and Performance
redis==5.2.1
prometheus-client==0.21.1

# WebSocket and async
websockets==14.1
python-multipart==0.0.18

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator>=2.1.0

# Identity Provider Integration (LDAP / AD)
ldap3>=2.9.1

# Logging
structlog==24.4.0

# NER / spaCy (PII detection model)
spacy==3.8.4

# Document extraction for attachment auditing
PyMuPDF==1.25.3
python-docx==1.1.2
openpyxl==3.1.5

# MITM Proxy for real traffic interception
mitmproxy==11.1.0

# MCP (Model Context Protocol) server for agent tool discovery
mcp>=1.26.0
